commit My version 1.2.0

+++ CAVmap/hotWordSearchModelData.m
@@ -46,7 +46,7 @@
     */
     
     
-    NSLog(@"%d",arr.count);
+    NSLog(@"%lu",(unsigned long)arr.count);
     
     
 }
